Latest Commit 19th November at 15:43

+++ PythonFiles/AdminPages/student_management.py
@@ -6,6 +6,7 @@ from flask_mail import Mail, Message
 from flask import Blueprint, render_template, session, request, redirect, url_for, flash
 from Authentication.middleware import auth
 
+
 managestudents_blueprint = Blueprint('managestudents', __name__)
 
 
@@ -19,14 +20,20 @@ def managestudents_auth(app):
 
     @managestudents_blueprint.route('/student_manage_dashbaord')
     def student_manage_dashbaord():
+        if not session.get('logged_in'):
+            # Redirect to the admin login page if the user is not logged in
+            return redirect(url_for('adminlogin.admin_login'))
         return render_template('AdminPages/StudentManagement/student_manage_dashbaord.html')
 
     @managestudents_blueprint.route('/admin_issue_dashboard', methods=['GET', 'POST'])
     def admin_issue_dashboard():
-        cnx = mysql.connector.connect(user='root', password='A9CALcsd7lc%7ac',
-                                      host=host,
-                                      database='ICSApplication')
-        cursor = cnx.cursor(dictionary=True)
+        if not session.get('logged_in'):
+            # Redirect to the admin login page if the user is not logged in
+            return redirect(url_for('adminlogin.admin_login'))
+        host = HostConfig.host
+        connect_param = ConnectParam(host)
+        cnx, cursor = connect_param.connect(use_dict=True)
+
         sql = """ SELECT * FROM application_page """
         cursor.execute(sql)
         result = cursor.fetchall()
@@ -39,10 +46,13 @@ def managestudents_auth(app):
 
     @managestudents_blueprint.route('/delete_student_management/<int:id>')
     def delete_student_management(id):
-        cnx = mysql.connector.connect(user='root', password='A9CALcsd7lc%7ac',
-                                      host=host,
-                                      database='ICSApplication')
-        cursor = cnx.cursor(dictionary=True)
+        if not session.get('logged_in'):
+            # Redirect to the admin login page if the user is not logged in
+            return redirect(url_for('adminlogin.admin_login'))
+        host = HostConfig.host
+        connect_param = ConnectParam(host)
+        cnx, cursor = connect_param.connect(use_dict=True)
+
         sql = "DELETE FROM application_page WHERE id = %s"
         cursor.execute(sql, (id,))
         # Close the cursor and database connection
@@ -52,10 +62,12 @@ def managestudents_auth(app):
 
     @managestudents_blueprint.route('/edit_student_admin_management/<int:id>', methods=['GET', 'POST'])
     def edit_student_admin_management(id):
-        cnx = mysql.connector.connect(user='root', password='A9CALcsd7lc%7ac',
-                                      host=host,
-                                      database='ICSApplication')
-        cursor = cnx.cursor(dictionary=True)
+        if not session.get('logged_in'):
+            # Redirect to the admin login page if the user is not logged in
+            return redirect(url_for('adminlogin.admin_login'))
+        host = HostConfig.host
+        connect_param = ConnectParam(host)
+        cnx, cursor = connect_param.connect(use_dict=True)
         sql = """SELECT * FROM application_page WHERE id = %s"""
         cursor.execute(sql, (id,))
         # Fetch all records matching the query
@@ -65,12 +77,15 @@ def managestudents_auth(app):
         cnx.close()
         return render_template('AdminPages/StudentManagement/edit_student_admin_management.html', records=records)
 
-    @managestudents_blueprint.route('/issue_resolve/<int:id>', methods=['GET', 'POST'])
-    def issue_resolve(id):  # -------------- VIEW STUDENT FORM
-        cnx = mysql.connector.connect(user='root', password='A9CALcsd7lc%7ac',
-                                      host=host,
-                                      database='ICSApplication')
-        cursor = cnx.cursor(dictionary=True)
+    @managestudents_blueprint.route('/viewStudentRecord/<int:id>', methods=['GET', 'POST'])
+    def viewStudentRecord(id):  # -------------- VIEW STUDENT FORM
+        if not session.get('logged_in'):
+            # Redirect to the admin login page if the user is not logged in
+            return redirect(url_for('adminlogin.admin_login'))
+        host = HostConfig.host
+        connect_param = ConnectParam(host)
+        cnx, cursor = connect_param.connect(use_dict=True)
+
         sql = """SELECT * FROM application_page WHERE id = %s"""
         cursor.execute(sql, (id,))
         # Fetch all records matching the query
@@ -78,14 +93,17 @@ def managestudents_auth(app):
         # Close the cursor and database connection
         cursor.close()
         cnx.close()
-        return render_template('AdminPages/StudentManagement/issue_resolve.html', records=records)
+        return render_template('AdminPages/StudentManagement/viewStudentRecord.html', records=records)
 
     @managestudents_blueprint.route('/old_user_insertion_by_admin', methods=['GET', 'POST'])
     def old_user_insertion_by_admin():
-        cnx = mysql.connector.connect(user='root', password='A9CALcsd7lc%7ac',
-                                      host=host,
-                                      database='ICSApplication')
-        cursor = cnx.cursor(dictionary=True)
+        if not session.get('logged_in'):
+            # Redirect to the admin login page if the user is not logged in
+            return redirect(url_for('adminlogin.admin_login'))
+        host = HostConfig.host
+        connect_param = ConnectParam(host)
+        cnx, cursor = connect_param.connect(use_dict=True)
+
         if request.method == 'POST':
             first_name = request.form['first_name']
             middle_name = request.form['middle_name']
@@ -149,13 +167,56 @@ def managestudents_auth(app):
 
     @managestudents_blueprint.route('/old_user_added_by_admin', methods=['GET', 'POST'])
     def old_user_added_by_admin():
-        cnx = mysql.connector.connect(user='root', password='A9CALcsd7lc%7ac',
-                                      host=host,
-                                      database='ICSApplication')
-        cursor = cnx.cursor(dictionary=True)
+        host = HostConfig.host
+        connect_param = ConnectParam(host)
+        cnx, cursor = connect_param.connect(use_dict=True)
+
         current_date = datetime.now().date()
         sql = """SELECT * FROM signup WHERE added_date = %s"""
         cursor.execute(sql, (current_date,))
         record = cursor.fetchall()
         cnx.commit()
         return render_template('AdminPages/StudentManagement/old_user_added_by_admin.html', record=record)
+
+    @managestudents_blueprint.route('/update_field/<int:id>/<field_name>', methods=['GET', 'POST'])
+    def update_field(id, field_name):
+        try:
+            cnx = mysql.connector.connect(user='root', password='A9CALcsd7lc%7ac',
+                                          host=host,
+                                          database='ICSApplication')
+            cursor = cnx.cursor(dictionary=True)
+            new_value = request.form['new_value']
+            # Build the SQL query to update the specified field value
+            sql = f"UPDATE application_page SET {field_name} = %s WHERE id = %s"
+            # Execute the SQL query with the specified field value and id
+            cursor.execute(sql, (new_value, id))
+            # Commit the transaction
+            cnx.commit()
+            # Close the cursor
+            cursor.close()
+            # Redirect the user to the edit_student_admin_management route with the corresponding id
+            return redirect(url_for('managestudents.edit_student_admin_management', id=id))
+        except Exception as e:
+            print("Error updating record:", e)
+            return None
+
+    @managestudents_blueprint.route('/delete_field/<int:id>/<field_value>', methods=['POST'])
+    def delete_field(id, field_value):
+        try:
+            cnx = mysql.connector.connect(user='root', password='A9CALcsd7lc%7ac',
+                                          host=host,
+                                          database='ICSApplication')
+            cursor = cnx.cursor(dictionary=True)
+            # Build the SQL query to delete records where email matches the specified value
+            sql = f"UPDATE application_page SET {field_value} = NULL WHERE id = %s"
+            # Execute the SQL query with the specified email value
+            cursor.execute(sql, (id,))
+            # Commit the transaction
+            cnx.commit()
+            # Close the cursor
+            cursor.close()
+            # Redirect the user to the edit_student_admin_management route with the corresponding id
+            return redirect(url_for('managestudents.edit_student_admin_management', id=id))
+        except Exception as e:
+            print("Error deleting record:", e)
+            return None
